requirements.txt
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
curl_cffi==0.11.4
frozendict==2.4.6
ibapi==9.81.1.post1
idna==3.10
lxml==6.0.0
multitasking==0.0.11
numpy==2.3.1
pandas==2.3.0
pandas-datareader==0.10.0
peewee==3.18.1
platformdirs==4.3.8
polygon-api-client==1.15.1
protobuf==6.31.1
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scipy==1.16.0
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
websockets==14.2
yfinance==0.2.64
